refactor(mapi-v2): separate portal resources
