+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/pom.xml
@@ -262,6 +262,15 @@
 							<inputSpec>${project.basedir}/src/main/resources/openapi/openapi-installation.yaml</inputSpec>
 						</configuration>
 					</execution>
+                    <execution>
+                        <id>model-portal</id>
+                        <goals>
+                            <goal>generate</goal>
+                        </goals>
+                        <configuration>
+                            <inputSpec>${project.basedir}/src/main/resources/openapi/openapi-portal.yaml</inputSpec>
+                        </configuration>
+                    </execution>
 					<execution>
 						<id>model-ui</id>
 						<goals>
