+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/resources/openapi/index-portal.html
@@ -1,0 +1,20 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+  <title>Management API v2 - UI</title>
+  <!-- Embed elements Elements via Web Component -->
+  <script src="https://unpkg.com/@stoplight/elements@7.7.17/web-components.min.js" integrity="sha512-er7CDLsLJUSwP4m6VU9pNH4bdjIrs3J/2DHLsbA3zCrWfGgvfSvftqmBQcqp4nHMafykJdMd4GbETNEIi34S+g==" crossorigin="anonymous"></script>
+  <link rel="stylesheet" href="https://unpkg.com/@stoplight/elements@7.7.17/styles.min.css" integrity="sha512-R0lX4nTGMLQSM9/IWxQD5cY/8e2VIcpcZNPaRuXdHjpCkxYxvmXF4uveEJk5NKqAGOximNajPCZIA5JwmE731w==" crossorigin="anonymous">
+</head>
+<body>
+
+<elements-api
+  apiDescriptionUrl="openapi-portal.yaml"
+  router="hash"
+  layout="sidebar"
+/>
+
+</body>
+</html>
