+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/resources/openapi/index.html
@@ -10,6 +10,7 @@
   <ul>
     <li><a href="index-apis.html">Gravitee.io APIM - APIs</a>
     <li><a href="index-plugins.html">Gravitee.io APIM - Plugins</a>
+    <li><a href="index-portal.html">Gravitee.io APIM - Portal</a>
     <li><a href="index-ui.html">Gravitee.io APIM - UI</a>
     <li><a href="index-installation.html">Gravitee.io APIM - Installation</a>
   </ul>
